backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymssql
python-dotenv
python-multipart
